+++ tests/Less/Lib/LessRulesetWithTestableRenderCss.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace ipl\Tests\Web\Less\Lib;
+
+use ipl\Tests\Web\Lib\TestableRenderCss;
+use ipl\Web\Less\LessRuleset;
+
+class LessRulesetWithTestableRenderCss extends LessRuleset
+{
+    use TestableRenderCss;
+}
