--- tests/Lib/LessRulesetWithTestableRenderCss.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace ipl\Tests\Web\Lib;
-
-use ipl\Web\LessRuleset;
-
-class LessRulesetWithTestableRenderCss extends LessRuleset
-{
-    use TestableRenderCss;
-}
